+++ remote-helper/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## 1.9.3
+
+-   Renamed to private-extension-marketplace
+
 ## 1.0.0
 
 -   Changes for marketplace release